incusd/network/openvswitch: Split OVN logic
Sponsored-by: Luizalabs (https://luizalabs.com)
